+++ ra/sources.py
@@ -10,10 +10,10 @@ def setup_sources(config_file):
     sources = [] # An array of empty souce objects
     config = load_cfg(config_file) # toml file
     for s in config['sources']:
-        coord = np.array(s['position'])
-        orientation = np.array(s['orientation'])
-        power_dB = np.array(s['power_dB'])
-        eq_dB = np.array(s['eq_dB'])
+        coord = np.array(s['position'], dtype=np.float32)
+        orientation = np.array(s['orientation'], dtype=np.float32)
+        power_dB = np.array(s['power_dB'], dtype=np.float32)
+        eq_dB = np.array(s['eq_dB'], dtype=np.float32)
         power_lin = 10.0e-12 * 10**((power_dB + eq_dB) / 10.0)
         delay = s['delay'] / 1000
         ################### cpp source class #################
