+++ ra_cpp/src/ptinpol.cpp
@@ -1,13 +1,13 @@
 #include "ptinpol.h"
 #include "isleft.h"
 
-int ptinpol(Eigen::Ref<Eigen::RowVectorXd> vert_x,
-        Eigen::Ref<Eigen::RowVectorXd> vert_y,
-        Eigen::Ref<Eigen::RowVector2d> ref_point2d)
+int ptinpol(Eigen::Ref<Eigen::RowVectorXf> vert_x,
+        Eigen::Ref<Eigen::RowVectorXf> vert_y,
+        Eigen::Ref<Eigen::RowVector2f> ref_point2d)
 {
 int Nvert = vert_x.cols()-1; // Get the number of rows
 int wn = 0; // start the winding number counter
-Eigen::RowVectorXd vy_refpy = vert_y.array() - ref_point2d(1);
+Eigen::RowVectorXf vy_refpy = vert_y.array() - ref_point2d(1);
 double isl = 0.0;
 
 for (int jv = 0; jv < Nvert; jv++) {
