+++ ra_cpp/src/raysphere.cpp
@@ -1,22 +1,19 @@
 #include "raysphere.h"
 
-double raysphere(Eigen::Ref<Eigen::RowVector3d> ray_origin, 
-    Eigen::Ref<Eigen::RowVector3d> v_dir,
-    Eigen::Ref<Eigen::RowVector3d> rec_coord,
+double raysphere(Eigen::Ref<Eigen::RowVector3f> ray_origin,
+    Eigen::Ref<Eigen::RowVector3f> v_dir,
+    Eigen::Ref<Eigen::RowVector3f> rec_coord,
     double rec_radius,
     double c0,
-    double dist_travel)
-{
+    double dist_travel){
     // Initialize timecross
-    double time_cross = 0.0;    
-
+    double time_cross = 0.0;
     // Calculate the roots for ray-sphere intersection
-    Eigen::RowVector3d ray_rec_vec = ray_origin - rec_coord;
+    Eigen::RowVector3f ray_rec_vec = ray_origin - rec_coord;
     double b = ray_rec_vec.dot(v_dir);
     double dist_origin2rec = ray_rec_vec.norm();
     double c = pow(dist_origin2rec, 2.0) - pow(rec_radius, 2.0);
     double delta = pow(b, 2.0) - c;
-
     // test of ray intersection
     if (delta >= 0.0 && b <= 0.0){
         // distance from center of receiver to mid crossing point
@@ -25,6 +22,5 @@ double raysphere(Eigen::Ref<Eigen::RowVector3d> ray_origin,
         double dL = sqrt(pow(dist_origin2rec, 2.0) + pow(d, 2.0));
         time_cross = (dist_travel + dL) / c0;
     }
-
     return time_cross;
 }
