+++ ra_cpp/src/refpoint.cpp
@@ -1,9 +1,9 @@
 #include "refpoint.h"
 
-Eigen::RowVector3d refpoint(Eigen::Ref<Eigen::RowVector3d> ray_origin,
-        Eigen::Ref<Eigen::RowVector3d> v_in,
-        Eigen::Ref<Eigen::RowVector3d> normal,
-        Eigen::Ref<Eigen::RowVector3d> vertcoord)
+Eigen::RowVector3f refpoint(Eigen::Ref<Eigen::RowVector3f> ray_origin,
+        Eigen::Ref<Eigen::RowVector3f> v_in,
+        Eigen::Ref<Eigen::RowVector3f> normal,
+        Eigen::Ref<Eigen::RowVector3f> vertcoord)
 {
     // Chech which normal of the plane to use
     double is_plane_detectable = v_in.dot(normal);
@@ -17,7 +17,7 @@ Eigen::RowVector3d refpoint(Eigen::Ref<Eigen::RowVector3d> ray_origin,
     // dot product of plane normal and one of plane vertexes
     double n_vert_dot = normal.dot(vertcoord);
     // calculate the reflection point
-    Eigen::RowVector3d ref_point = ray_origin +
+    Eigen::RowVector3f ref_point = ray_origin +
         ((n_vert_dot - n_ro_dot) / is_plane_detectable) * v_in;
     return ref_point;
 }
